Refactor and fix results
