+++ source/liblocate/CMakeLists.txt
@@ -37,7 +37,6 @@ set(headers
 
 set(sources
     ${source_path}/liblocate.c
-    ${source_path}/liblocate.cpp
     ${source_path}/utils.c
 )
