+++ source/liblocate/include/liblocate/liblocate.h
@@ -123,27 +123,3 @@ LIBLOCATE_API void locatePath(char ** path, unsigned int * pathLength, const cha
 #ifdef __cplusplus
 }
 #endif
-
-
-#ifdef __cplusplus
-
-
-namespace liblocate
-{
-
-
-LIBLOCATE_API std::string getExecutablePath();
-
-LIBLOCATE_API std::string getBundlePath();
-
-LIBLOCATE_API std::string getModulePath();
-
-LIBLOCATE_API std::string getLibraryPath(void * symbol);
-
-LIBLOCATE_API std::string locatePath(const std::string & relPath, const std::string & systemDir, void * symbol);
-
-
-} // namespace liblocate
-
-
-#endif
