+++ source/tests/liblocate-test/liblocate_test.cpp
@@ -34,14 +34,6 @@ TEST_F(liblocate_test, getExecutablePath_Return)
     free(executablePath);
 }
 
-TEST_F(liblocate_test, getExecutablePath_Return_CPP)
-{
-    const auto result = liblocate::getExecutablePath();
-
-    EXPECT_LT(0, result.size());
-    EXPECT_FALSE(result.c_str() == 0x0);
-}
-
 TEST_F(liblocate_test, getBundlePath_NoReturn)
 {
     getBundlePath(nullptr, nullptr);
@@ -63,18 +55,6 @@ TEST_F(liblocate_test, getBundlePath_Return)
     EXPECT_EQ(length, strlen(bundlePath));
 
     free(bundlePath);
-#else
-    SUCCEED();
-#endif
-}
-
-TEST_F(liblocate_test, getBundlePath_Return_CPP)
-{
-#ifdef SYSTEM_DARWIN
-    const auto result = liblocate::getBundlePath();
-
-    EXPECT_LT(0, result.size());
-    EXPECT_NE(0x0, result.c_str());
 #else
     SUCCEED();
 #endif
@@ -102,14 +82,6 @@ TEST_F(liblocate_test, getModulePath_Return)
     free(modulePath);
 }
 
-TEST_F(liblocate_test, getModulePath_Return_CPP)
-{
-    const auto result = liblocate::getModulePath();
-
-    EXPECT_LT(0, result.size());
-    EXPECT_FALSE(result.c_str() == 0x0);
-}
-
 TEST_F(liblocate_test, getLibraryPath_NoReturn)
 {
     getLibraryPath(reinterpret_cast<void*>(getExecutablePath), nullptr, nullptr);
@@ -130,14 +102,6 @@ TEST_F(liblocate_test, getLibraryPath_Return)
     EXPECT_EQ(length, strlen(libraryPath));
 
     free(libraryPath);
-}
-
-TEST_F(liblocate_test, getLibraryPath_Return_CPP)
-{
-    const auto result = liblocate::getLibraryPath(reinterpret_cast<void*>(getExecutablePath));
-
-    EXPECT_LT(0, result.size());
-    EXPECT_NE(nullptr, result.c_str());
 }
 
 TEST_F(liblocate_test, locatePath_NoReturn)
@@ -173,17 +137,6 @@ TEST_F(liblocate_test, locatePath_Return)
     EXPECT_EQ(length, strlen(path));
 
     free(path);
-}
-
-TEST_F(liblocate_test, locatePath_Return_CPP)
-{
-    const auto relPath = std::string("source/version.h.in");
-    const auto systemPath = std::string("share/liblocate");
-
-    const auto result = liblocate::locatePath(relPath, systemPath, reinterpret_cast<void*>(getExecutablePath));
-
-    EXPECT_LT(0, result.size());
-    EXPECT_NE(nullptr, result.c_str());
 }
 
 TEST_F(liblocate_test, locatePath_ReturnNoSymbol)
